% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_boot.R
\name{var_boot}
\alias{var_boot}
\title{Calculate the naive nonparametric bootstrap variance estimator for AUC}
\usage{
var_boot(formula_string, label_true, data, B, link = "logit")
}
\arguments{
\item{formula_string}{A string with an expression of the form \code{y ~ model} that represents
the binary classification model. It may include operators as +, ^ and :}

\item{label_true}{A vector of the true labels in the data set, coded as 1 (positive) and 0 (negative)}

\item{data}{A data frame, list or environment containing the variables in the model.
It can also be an object coercible by as.data.frame
to a data frame.}

\item{B}{An integer indicating the desired number of bootstrap samples}

\item{link}{A string specifying the model link function for glm function
used to fit the binomial model. Possible links are \code{logit}., \code{probit}, \code{cauchit},
(corresponding to logistic, normal and Cauchy CDFs respectively) \code{log} and
\code{cloglog} (complementary log-log). The default is \code{logit}.}
}
\value{
The value of the bootstrap variance estimator for the AUC.
}
\description{
Calculate the naive nonparametric bootstrap variance estimator for AUC
}
\examples{
library(aucvar)
my_data <- na.omit(breastcancer) # Omit NA values
model_formula <- "Class~`Clump Thickness`+ `Uniformity of Cell Size`+`Uniformity of Cell Shape`+
`Marginal Adhesion` + `Single Epithelial Cell Size` + `Bare Nuclei` +
`Bland Chromatin` + `Normal Nucleoli` + `Mitoses`"
# Use quotes inside double quotes since data set variable names have spaces
var_boot(model_formula, my_data$Class, my_data, B = 10^3)

}
\references{
\cite{B. Efron and C. Stein (1981). The jackknife estimate of variance. The Annals of
Statistics 9: 586-596.}
}
