% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_jackknife.R
\name{var_jackknife}
\alias{var_jackknife}
\title{Calculate the jackknife variance estimator for AUC}
\usage{
var_jackknife(formula, label_true, data, B = Inf, d)
}
\arguments{
\item{formula}{A string with an expression of the form y ~ model that represents
the binary classification model. It may include operators as +, ^ and :}

\item{label_true}{A vector of the true labels in the data set, coded as 1 (positive) and 0 (negative)}

\item{data}{A data frame, list or environment containing the variables in the model
except for the response variable. It can also be an object coercible by as.data.frame
to a data frame.}

\item{B}{An integer indicating the desired number of jackknife samples}

\item{d}{Number of data entries to remove to generate the jackknife samples}
}
\value{
The value of the jackknife variance estimator for the AUC
}
\description{
Calculate the jackknife variance estimator for AUC
}
\examples{
'TODO'
}
