% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Calculate the area under the ROC curve}
\usage{
auc(p_pred, label_true)
}
\arguments{
\item{p_pred}{A vector of the predicted probabilities for the observations
in the data set.}

\item{label_true}{A vector containing the true class labels in the dataset.}
}
\value{
Area under the ROC curve, which is equal to the Mann-Whitney
two-sample U-statistic. It is also the estimated probability that the binary classifier will score a
randomly drawn positive sample higher than a randomly drawn negative sample.
}
\description{
Calculate the area under the ROC curve
}
\examples{
library(aucvar)
mydata <- na.omit(breastcancer) # Omit NA values
optimal_model <- glm(Class~`Clump Thickness`+`Uniformity of Cell Shape`+
`Bare Nuclei` + `Bland Chromatin`, family=binomial(link="logit"), data=mydata)
prob <- predict(optimal_model, type="response")
labels <- mydata$Class
auc(prob, labels)
}
