% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_AUC.R
\name{CI_AUC}
\alias{CI_AUC}
\title{Asymptotic confidence interval of AUC based on an unbiased variance estimator
and the asymptotic normality}
\usage{
CI_AUC(p_pred, label_true, conf_level = 0.95)
}
\arguments{
\item{p_pred}{A vector of the predicted probabilities for the observations in the data set}

\item{label_true}{A vector of the true labels in the dataset, coded as 1 (positive) and 0 (negative)}

\item{conf_level}{The confidence level required.The default is 0.95.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits
for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in \%
(by default 2.5\% and 97.5\%).
}
\description{
Asymptotic confidence interval of AUC based on an unbiased variance estimator
and the asymptotic normality
}
\examples{
'TODO'
}
