% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varAUC.R
\name{varAUC}
\alias{varAUC}
\title{Calculate an unbiased variance estimator of AUC}
\usage{
varAUC(p_pred, label_true, B = Inf)
}
\arguments{
\item{p_pred}{A vector of the predicted probabilities for the observations in the data set}

\item{label_true}{A vector of the true labels in the dataset, coded as 1 (positive) and 0 (negative)}

\item{B}{The number of random partitions to use in the partition-resampling scheme. If B is
set to Inf or it is ommited, the exact number of possible partitions will be calculated.}
}
\value{
The value of the AUC variance estimator
}
\description{
Calculate an unbiased variance estimator of AUC
}
\examples{
library(aucvar)
mydata <- na.omit(breastcancer) # Omit NA values
optimal_model <- glm(Class~`Clump Thickness`+`Uniformity of Cell Shape`+
`Bare Nuclei` + `Bland Chromatin`, family=binomial(link="logit"), data=mydata)
predictions <- predict(optimal_model, type="response")
varAUC(predictions, mydata$Class, 10^3)
}
