% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varUn.R
\name{varUn}
\alias{varUn}
\title{Compute the variance of a general 2-sample U-statistic}
\usage{
varUn(sample1, sample2, k1, k2, phi, B = Inf)
}
\arguments{
\item{sample1}{The first sample.}

\item{sample2}{The second sample.}

\item{k1}{The number of observations from sample 1.}

\item{k2}{The number of observations from sample 2.}

\item{phi}{The kernel function for the U-statistic}

\item{B}{The number of random partitions.}
}
\value{
The variance of the 2-sample U-statistic
}
\description{
Compute the variance of a general 2-sample U-statistic
}
\examples{
'TODO'
}
